refactor: rename methods responses in response
