+++ src/Cancel/Response.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace DansMaCulotte\Monetico\Cancel;
+use DansMaCulotte\Monetico\Recovery\Response as RecoveryResponse;
+
+
+class Response extends RecoveryResponse
+{
+
+}
