--- src/Recovery/CancelResponse.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace DansMaCulotte\Monetico\Recovery;
-
-
-class CancelResponse extends RecoveryResponse
-{
-
-}
