+++ tests/CancelTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use Carbon\Carbon;
-use DansMaCulotte\Monetico\Recovery\Cancel;
+use DansMaCulotte\Monetico\Cancel\Cancel;
 use PHPUnit\Framework\TestCase;
 use DansMaCulotte\Monetico\Exceptions\Exception;
